Move WaylandContext in its own file.
